updated by kamalakar

+++ partyanalyst-service/trunk/src/main/java/com/itgrids/partyanalyst/notification/service/impl/Scheduler.java
@@ -2,6 +2,7 @@ package com.itgrids.partyanalyst.notification.service.impl;
 import com.itgrids.partyanalyst.notification.service.ISchedulerService;
 import com.itgrids.partyanalyst.service.IMailsSendingService;
 import com.itgrids.partyanalyst.utils.DateUtilService;
+import com.itgrids.partyanalyst.utils.IConstants;
 
 public class Scheduler {
 	
@@ -27,11 +28,17 @@ public class Scheduler {
 	
 	public void runTheBatchJobForEveryDay()
 	{
+		if(!IConstants.DEFAULT_SCHEDULER_SEVER.equalsIgnoreCase(IConstants.SERVER))
+			return;
+		
 		schedulerService.deleteSearchEngineAccessedURLsFromUserTracking(dateUtilService.getCurrentDateAndTime(),dateUtilService.getCurrentDateAndTime());
 	}
 	
 	public void runTheBatchJobForEveryWeek()
 	{
+		if(!IConstants.DEFAULT_SCHEDULER_SEVER.equalsIgnoreCase(IConstants.SERVER))
+			return;
+		
 		mailsSendingService.sendMailsToPasswordnotUpdatedusers();
 	}
 }
